refactor: move Notifier implementations from engine to app target
The engine library was importing AppKit and UserNotifications via
the two concrete Notifier implementations, violating the "engine is
process-clean" rule (the engine should be free of UI-framework
dependencies so it stays embeddable from the Camera Extension's
sandbox and from headless test harnesses).

Calibrated 82/100 in the engine slop review.

Changes:

- Sources/AVPainReliever/Adapters/Notifier.swift now holds only the
  protocol + the two convenience extension methods. Doc comment
  trimmed to be implementation-agnostic.
- Sources/AVPainRelieverApp/Notifiers.swift (new) holds
  UserNotificationsNotifier + AppleScriptNotifier verbatim. The
  `#if canImport(UserNotifications)` / `#if canImport(AppKit)`
  guards drop away because the app target always has these.
- AppDelegate's selection logic (UserNotificationsNotifier vs
  AppleScriptNotifier based on Bundle.main.bundleIdentifier) is
  unchanged — the names resolve to the new file via the same
  target.

Engine target now imports zero UI frameworks
(`grep -rn "import AppKit\\|import UserNotifications\\|import SwiftUI"
Sources/AVPainReliever/` → empty).

Behavior unchanged. swift test: 183 passed.
